Move the search, framework and revision dropdowns, and the download button, closer to the table
